Implement bug fix. See below. Bug: `ShoukakuPlayer#track` is incorrectly null.
Cause:
`ShoukakuPlayer#playTrack(track, { noReplace: false })`

When `ShoukakuPlayer#playTrack(track, { noReplace: false }` is called, `ShoukakuPlayer#track` is set correctly to the new track. However, the following `end` event sets it to null. The `end` event occurs because of `{ noReplace: false }` provided in the function.

Fix:
Listen to the `start` event and set `ShoukakuPlayer#track` to the track stored in the event data. This will also help with keeping Shoukaku and LavaLink synced.